research/claude_code/pyproject.toml
[project]
name = "claude-code-research"
version = "0.1.0"
description = "Research project exploring agentic AI coding assistant patterns."
dependencies = [
    "jinja2>=3.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "google-genai>=0.1.0",
    "requests>=2.31",
]

[tool.uv]
managed = true
